# perf: use a sync pool for machine constructor