--- a/pkgs/gnolang/machine.go
+++ b/pkgs/gnolang/machine.go
@@ -137,11 +137,15 @@ func NewMachineWithOptions(opts MachineOptions) *Machine {
 	return mm
 }
 
-const OpSize = 1024
-const ValueSize = 1024
+const (
+	OpSize    = 1024
+	ValueSize = 1024
+)
 
-var opZeroed [OpSize]Op
-var valueZeroed [ValueSize]TypedValue
+var (
+	opZeroed    [OpSize]Op
+	valueZeroed [ValueSize]TypedValue
+)
 
 // m should not be used after this call
 // if m is nil, this will panic
@@ -149,7 +153,7 @@ var valueZeroed [ValueSize]TypedValue
 // and prevent objects that were not taken from
 // the pool, to call Release
 func (m *Machine) Release() {
-	//copy()
+	// copy()
 	// here we zero in the values for the next user
 	m.NumOps = 0
 	m.NumValues = 0