This PR uses a `sync.Pool` in the `Machine` constructor.
This improves the following
1. Reduces allocations and deallocations
2. Improves runtime and memory consumption
3. Because of less objects in memory, less GC pauses

[issue](https://github.com/gnolang/gno/issues/624) with benchmarks from master.

Benchmarks on this branch.

CPU bench
```
Type: cpu
      10ms   100%   100%       10ms   100%  github.com/gnolang/gno/pkgs/gnolang.(*Machine).PopAsPointer
         0     0%   100%       10ms   100%  github.com/gnolang/gno/pkgs/gnolang.(*Machine).Run
         0     0%   100%       10ms   100%  github.com/gnolang/gno/pkgs/gnolang.(*Machine).RunMain
         0     0%   100%       10ms   100%  github.com/gnolang/gno/pkgs/gnolang.(*Machine).RunStatement
         0     0%   100%       10ms   100%  github.com/gnolang/gno/pkgs/gnolang.(*Machine).doOpAddAssign
         0     0%   100%       10ms   100%  github.com/gnolang/gno/pkgs/gnolang.BenchmarkBlah
         0     0%   100%       10ms   100%  testing.(*B).launch
         0     0%   100%       10ms   100%  testing.(*B).runN

```

Memory profile

```
Showing nodes accounting for 10958.29kB, 72.79% of 15054.80kB total
Showing top 10 nodes out of 112
      flat  flat%   sum%        cum   cum%
 4097.83kB 27.22% 27.22%  4097.83kB 27.22%  github.com/gnolang/gno/pkgs/gnolang.addAssign
 1537.50kB 10.21% 37.43%  1537.50kB 10.21%  runtime.allocm
 1184.27kB  7.87% 45.30%  1184.27kB  7.87%  runtime/pprof.StartCPUProfile
  902.59kB  6.00% 51.29%  1553.21kB 10.32%  compress/flate.NewWriter
  650.62kB  4.32% 55.62%   650.62kB  4.32%  compress/flate.(*compressor).init
  528.17kB  3.51% 59.12%   528.17kB  3.51%  regexp.(*bitState).reset
  519.03kB  3.45% 62.57%   519.03kB  3.45%  time.init
  513.50kB  3.41% 65.98%   513.50kB  3.41%  github.com/cockroachdb/apd.(*Decimal).setString
  512.50kB  3.40% 69.39%   512.50kB  3.40%  github.com/gnolang/gno/pkgs/amino.(*Codec).parseStructInfoWLocked
  512.28kB  3.40% 72.79%  1024.29kB  6.80%  github.com/gnolang/gno/pkgs/gnolang.(*defaultStore).SetCacheType

```
As we can see, we no longer have the `Machine` constructor in the top 10 of either.
And we can see that the memory consumption went from 28666.37kB to 15054.80kB.