@peter7891 2023-03-21
> Nice. What was used for benchmarking?
> 
> Added one comment for feedback above.

`go tool pprof`