@moul 2023-03-28
@jaekwon, do you think @peter7891's answer is satisfactory? Can we merge his work? Should we add any comments or issues for future improvements, or include unit tests?